@@ -1,10 +1,19 @@
 import find from 'lodash-es/find';
 import includes from 'lodash-es/includes';
-import { Observable } from 'rxjs/Observable';
+import { Observable } from 'rxjs';
 import { JsonApiDatastore, ModelType } from '../services/json-api-datastore.service';
 import { ModelConfig } from '../interfaces/model-config.interface';
 import * as _ from 'lodash';
 import { AttributeMetadata } from '../constants/symbols';
+
+/**
+ * HACK/FIXME:
+ * Type 'symbol' cannot be used as an index type.
+ * TypeScript 2.9.x
+ * See https://github.com/Microsoft/TypeScript/issues/24587.
+ */
+// tslint:disable-next-line:variable-name
+const AttributeMetadataIndex: string = AttributeMetadata as any;
 
 export class JsonApiModel {
   data: any;
@@ -43,12 +52,12 @@ export class JsonApiModel {
   }
 
   save(params?: any, headers?: Headers): Observable<this> {
-    const attributesMetadata: any = this[AttributeMetadata];
+    const attributesMetadata: any = this[AttributeMetadataIndex];
     return this._datastore.saveRecord(attributesMetadata, this, params, headers);
   }
 
   get hasDirtyAttributes() {
-    const attributesMetadata: any = this[AttributeMetadata];
+    const attributesMetadata: any = this[AttributeMetadataIndex];
     let hasDirtyAttributes = false;
     for (const propertyName in attributesMetadata) {
       if (attributesMetadata.hasOwnProperty(propertyName)) {
@@ -64,7 +73,7 @@ export class JsonApiModel {
   }
 
   rollbackAttributes(): void {
-    const attributesMetadata: any = this[AttributeMetadata];
+    const attributesMetadata: any = this[AttributeMetadataIndex];
     let metadata: any;
     for (const propertyName in attributesMetadata) {
       if (attributesMetadata.hasOwnProperty(propertyName)) {
@@ -80,7 +89,7 @@ export class JsonApiModel {
       }
     }
 
-    this[AttributeMetadata] = attributesMetadata;
+    this[AttributeMetadataIndex] = attributesMetadata;
   }
 
   get modelConfig(): ModelConfig {
